Added a GVT manager code to compute GVT  in the system. The gvt algorithm is based on Mattern's GVT algorithm. Including the GVT manager into the simulation required several changes to the Simulation and Communicator class.
In addition, added a method to the scheduler to help compute LGVT value.

Added a color instance variable to the event to track color of events for
GVT calculation.

The check-in also incldues some general code formatting and code clean-up 
operations as well. A lot of the code is not even indented properly. 
In some of the files (the ones I did modify much), the code is nasty 
and I need to clean it up some time soon.