Some more example clean-up of Scout event in BugLife simulation to illustrate example of how to deveop a C++ class with suitable documentation.
In addition revised the API of Scout event to exclude the event type 
parameter because the event type for Scout event is always fixed to be
"SCOUT".  Passing the same parameter in each call is rather silly, 
particularly when the event type should be actually hard coded for 
correctness.

In addition, made the event type in the BugEvent base class to be constant
to ensure that it does not changes.

More importantly converted the getEventSize() method to be a constant
method to be consistent and compatible with MUSE API requirements.

Similar changes to all the event class needs to happen before the simulation
successfully runs in parallel.