Adding a custom SSH library (based on libSSH2) to ease using SSH in Qt GUI. Unfortunatley, the library is currently undocumented but its operation has been reasonably tested.  The SSH library is in `infra/ssh` directory.
Added a custom File Dialog (in core/filediag) to ease seamless operation
between local and remote machines. The custom file dialog utilises the
SSH library (in `infra/ssh` directory) for accessing remote servers. 
Ideally, the custom file dialog should be consistently used in MUSE-GUI 
in lieu of `QFileDialog`. The API of `QFileDialog` needs to be expanded
to mirror some (not all) of the features provided by `QFileDialog' to
ease accessing the user's selection. 

The custom file dialog can be invoked in the following manner:

    QStringList filters;
    filters << "Image files (*.png *.xpm *.jpg)"
            << "Text files (*.txt)"
            << "Any files (*.*)"
            << "Directories only";

    CustomFileDialog dialog(NULL, QDir::homePath());
    dialog.setFileFilters(filters);
    dialog.exec();