The Load_GIS.ecl file is now a module that contains data. When a filename is passed to the module, two embedded C++ functions are executed. The first is as before, it extracts the lat/lon data into a record format. The second extracts any metadata into a different record format. Both records can be accessed through scoping of the module.
Work on writing existing GIS data in record format to a newly created GIS file has started. The new creation of a GIS file is why the metatata is important as it can be tedious to come up with it by hand. It is almost done, should be done by the end of the day.