Revised and rebuilt the framework on how GIS files are loaded in as records then written back as GIS files.
NOTE: Framework is still in the midst of being developed; this commit just serves as a mid-step towards the completion. 

------------------------------------------------------------------------

Previously, for loading a GIS file, there was a chunk of C++ code that used GDAL to read a GIS file then passed it back as a Dataset. This was a buggy way of reading the GIS files, though, because the GIS files only existed on the server with the dropzone. As a result, the servers that did not have access to the GIS file crashed -- this is the initial cause of why the Dataset could not be distributed after being loaded because most of the nodes became offline.  

The new framework for loading a GIS file in as a Dataset now is made up of a few more steps. 
1) C++ code uses GDAL to read the GIS file and create a temporary BLOB file. The BLOB file is structured as latitude (4 bytes) longitude (4 bytes) value (4 bytes). 
2) The STD ECL library is used to spray the BLOB as a fixed record size of 12 bytes. The BLOB is then deleted. 
3) The logical file created a a result of spraying is used to create the Dataset. 

Given the above of how a GIS file is loaded in as a Dataset, a GIS file is written by going through the series of steps in reverse order. 

-------------------------------------------------------------------------

What is completed so far:
1) The Steps 1,2,3 are fleshed out for the loading of a GIS file. 

Problems encountered
1) Steps 1 and 2 need to be executed as separate work units. 

What needs to be done
1) Finish debugging the loading of GIS data
2) Develop the code for writing of GIS data