Minor fix in Scheduler.cpp to remove an agent from the map.
Major fix in Simulation.cpp to reposition an migrated agent in the scheduler's
heap, because after migration, its LVT is no longer the same/meanigful.

With these changes PHold seems to run all the way to the end, but it does not
terminate due to some logic issue  in Simulation.cpp at this line:

           MPI::COMM_WORLD.Allgather(&doSwap, 1, MPI::BOOL, 
                                     swapArray, 1, MPI::BOOL);

It is unclear what that line is trying to accomplish. However, that may need to
be appropriately fixed.