Experimentation indicated that the general "distance" of nodes from the root was providing a misleading (and smaller) average measure due to the number of samples recorded during simulation.
For example, the EpiFlu tree for Turkey has 91 nodes in 2
clusters. The average intra-cluster distance is about 0.9 nt with
several nodes having zero distance.  The Simulation has about 300
nodes with many nodes having zero distance causing the average to be
much smaller at 0.1 showing it as ~90% error which is misleading.

I expanded the tree-comparison scripts to:

1. Create clusters based on phylogenetic distances (which are now
   reported correctly by the simulator)

2. Report the number of clusters in simulated vs. EpuFlu (actual)
   phylograms.

3. Report inter-clade (between clades) distances and intra-clade
   distances separately. Now errors in inter-clade distances are
   pretty small.  Intra-clade distance percentage values are high but
   the averages are < 1 nt (which is essentially noise)

With these changes the scripts provide more comprehensive statistics
for comparisons.