Fixed a reasonably conspicuous bug in schedule generation. The one line bug was at ScheduleGenerator.cpp:2426
The code was incorrectly accessing `transportRadiusLimits[2]`
(undefined value for this array) instead of
`transportRadiusLimits[1]`.  Note that `transportRadiusLimits[1]`
corresponds to `Max_Walking_Distance` in the configuration file.

This fixes an inconsistent access issue in Haplos to make it move
further.  However, this is not a full/permanent fix.  The full fix
would be:

1. Change `transportRadiusLimits` to a vector of values to avoid
   issues with knowing what the bounds of this pointer are.

2. Change magic numbers like '0' and '1' to enumerations or
   pre-defined named-constants such as: 'MaxWalkDistance',
   'MaxPubTranspDist' etc.

A minor new-line missing fix in main.cpp