With the previous check-in of changing std::list to std::deque caused parallel simulations to fail.  The source of the issue was an incorrect assumption about how iterators are to be handled when erasing events in input and output queues during rollback recovery.
The previous version of this code was using a pair of iterators to
remove events -- one iterator was advanced to point to the next
element and the previous element was erased.  This logic works when
using std::list with uses an linked list and the iterator to the
next-element is not invalidated.

However, the above logic does not work with std::deque or std::vector
as erasing a previous element moves/shifts elements in the queue.

The solution is quiet simple -- use the API correctly -- by writing
erase logic as:

   iter = container.erase(iter);

where the erase method returns the correct iterator pointing to the
next element (or end() if it is last element in the list) in the
container. Accordingly, this commit rewrites the logic in
doCancellationPhaseOutputQueue doCancellationPhaseInputQueue to use
iterators/erase() method correctly.

This fixes runtime issues and the parallel simulation operates
correctly.