One surprising outcome from running tests using 100,000 agents was that the overhead of creating Tier2Entry objects is pretty high.  Initially, with 100k agents, 2tHeap was almost 2x faster than heap2tQ (btw, this queue should become 3tHeap instead) with 20 concurrent events, which it should not be.  However, it was because of the overhead of creating new Tier2Entry objects, specifically, the vector-of-event-pointers encapsulated by it.
In order to reduce the overhead of object creation, the heap2tQ now
reuses/recycles Tier2Entry. In order to eliminate conflict with other
scheduler-queues, a new HOETier2Entry class has been added.  In addtion,
the second tier list now contains pointers to HOETier2Entry objects rather
than objects themselves. This streamlines the process of recycling the
objects (via pointers). The scheduler-queue also includes a list of entries
that can be recycled. 

Hopefully the recycling of HOETier2Entry objects will help address the
surprising/counterintuitive observation in 100k agents.  Due to the changes
the data for sequential simulations needs to be regenerated.  On to that
next.