Missed checking-in the friend clause additions for TwoTierLadderQueue to Event.h
Fixed a minor bug in erasing events in bottom which is sorted -- since
it is sorted the "swap with end and pop_back" optimization cannot be
used here.  Have to erase in place.

With this change PHOLD runs in parallel with 2tLadderQ.  However, the
parallel runs with rollbacks are not as fast as heap2tQ. Need to
profile the 2tLadderQ code to see where the problem lies.