Corrected collection of statistics for average events scanned in top & rungs for 2-tier heap -- this queue only compares events in a sub-bucket. So the statistics object is passed down to TwoTierBucket to correctly track the number of events being scanned.
Added a counter called `cdNoCanScanBot` to track scans in bottom where
no cancellations actually happened.

This change is consistent with the changes in LadderQueue ---

Previously, erase_after method accepted a flag to distinguish sorted
(for bottom only) and unsorted buckets (top & rungs) so that for
unsorted lists, events can be quickly removed by swapping with back
and pop_back.

However, while profiling 2-tier ladder queue, I noticed that the if we
split up the event cancellation for bottom (sorted list) and buckets
(unsorted list) into 2 separate methods, the queue seemed to
experience fewer rollbacks and consequently performed more reliably.
This is not very consistent/reproducible on different settings.

However, to avoid potential issues, I have restructured the code and
TwoTierBottom impelements the erase_after for the sorted list of
events.