This check-in essentially addresses 1 scenario that was encountered with Ladder queue (and also affects 2-tier ladder queue) -- under heavy rollback, the bottom would get very large -- almost have 500,000 events with 10k agents (delay 0, eventsPerAgent=3), typically only on 1 MPI-process.
Once the bottom gets this large, the whole simulation essentially
slows down to a crawl and the simulation takes for ever to finish.

Having studied this situation for well over a day, I have made the
following changes to ladder queue (and these changes need to be
implemented in 2-tier ladder queue as well):

1. Removed the limit on the number of rungs of the ladder when
   moving bottom to ladder.

2. Movement of bottom to ladder happens only when the bottom has
   events that are not at the same time stamp.  If bottom has all
   events at the same time, there is no point in moving events.

3. Revised the rung-clean up operation in the ladder queue to occur
   outside 'recurseRung' method.  This clean-up is closer to the
   paper's version and seems to keep the ladder depth in better check.

4. Added a counter to track the max bottom length for validation.

5. Added a check to generate an error and abort if bottom length
   exceeds 50,000 events. The 50k is an arbitrary number.

These changes need to be incorporated into 2-tier ladder queue,
specifically, the part of moving bottom into the ladder.