Now that the overall troubleshooting of the ladder queue implementations are almost all settled, I updated the implementation involving heap-bottom and multi_set-bottom to operate consistently & correctly with the updates.
In addition, to streamline implementation of the different bottom-rung
data structures, I modified the ListBucket to use std::list (a
doubly-linked list) instead of std::forward_list (a singly-linked
list).  The change from std::forward_list to std::list did not make
any difference in performance of ListBucket, which continues to
perform slower than VectorBucket.

Completing the implementation of the different data structures to
operate consistently will enable streamlining experiments for
comparing different options for developing ladder queue.  The code
however is not documented and that needs to happen prior to paper
submission.