Added raw data GSA analysis for parallel simulation configuration. The charts from this data need to be plotted next.
Although I started these runs on Tuesday, the Redhawk cluster had issues and 
continues to have issues running jobs correctly in parallel. So jobs get suck
in the job queue for 12 hours or more and have to be manually resubmitted to
keep things working. Furthermore, output from the jobs is not saved into 
.o files and is simply lost. 

This issue started after Jan 2 2017 maintenance on the cluster. IT folks are 
aware of this issue and mentioned they will look into it. However, after 3 
days there is still no resolution.