Earlier I commited a change to avoid adding events to input & output queue of agents in *sequential* (or 1 process) mode to reduce uncessary operations.
However, I had missed to decrease reference on these events causing the 
simulation not to delete the events during garbage collection. This was
causing memory to grow uncontrollably in cases were RECYCLE_EVENTS was not
used.  

This check-in fixes this issue by decreasing the reference count for events
(for both input and output queues) in sequential (or 1 process) simulations.