General clean-up of the code.  One big change is adding documentation to the various instance variables and methods based on the description in Craother's paper.
In general I am chaning "unsigned int" to "muse::Time" so that time is
consistently used throughout the model.

During these changes I noticed that properties such as "channel" (this
is just one example, there are many like this one throughout the code)
is incorrectly used -- instead of state variables being used, instance
variables in "PCSAgent" class are incorrectly being used.  This is a
serious defect with the PCSAgent and I will be fixing it next.