Moved sessions from default storage in disk files to store into a database. I tested the migration on the development MySQL database on hadoop1.csi.miamioh.edu via the command:
php artisan migrate --env=development

and it worked fine. I checked and the tables were created. I also noticed 
that the 'analyses', 'models', and 'users' tables were also successfuly 
created in MySQL.

I did not notice the 'groups' table being created. Maybe we need a groups
table to just list the set of valid groups we have in the system?