Added initial plumbing for loading JSON format of metadata for a model into the database. Currently the PHP code uses a C++ binary helper to extract JSON formatted data from a binary model file.
The idea of storing the JSON in the database is to ease serving it 
out to clients.

Current check-in only loading metadata for model.  The next step is 
load the actual model data in longitudinal (based on time for each entity
in the model to display line charts per entity) and transverse (all entities
at a given timestamp to display model state on a given day)