Some more testing to ensure the Java and C++ side are operating in the same manner (operations are getting close, still need to test event cancellation features) --
Java Bug fix: Fixed bug with upper_bound which was not implemented
correctly earlier.  This was causing some events to be scheduled
incorrectly in the Java version.

Included some (commented-out by default) output statements in the C++
and Java benchmarks to print events and schedules to ease comparing
the two versions.