Reviewed "Introduction" chapter -- that looks fine.
The "Related works" must be a chapter in a thesis. This is an
important chapter to demonstrate that the student has done literature
review and summarized relevant papers.  Given that data structures are
widely researched, this chapter is very weak with just few papers.

Reworked the "related works" chapter with some more verbiage.  Moved
ladder queue description from chapter 5 to related work because it
seems to fit better in this "related works" chapter.  There are 2
figures in this chapter that need to be redrawn to avoid copyright
issues.