Cleaned-up and simplified the implementation of 3-tier heap to be consistent with the data structure implementation in C++.
The earlier implementation was using a mishmash of sub-structures
(like Fibonacci heap etc.) which was incorrect implementation of
3-tier heap.  Now the 3-tier heap is correctly implemented as a binary
heap backed by an ArrayList (which is what the C++ implementation does
to ensure very low constants for its operations).

Short runs seem to be operating ok. But more testing is needed to
ensure the 3tHeap is operating correctly.