Big UI update around groups.
Replaced the concept of "role" with that of a "group". Overhauled user management dashboard to include all UI components for managing roles, however backend for adding a role to a user and deleting a role is not yet implemented.

Also started on a framework for asynchronous updates by easily calling a single method from anywhere in javascript to initiate an ajax call to update value. Should make future development that involves editing data base values easier.

Note: if you've migrated since my last update, you may have issues migrating as I changed the names. If this is an issue, I can instead make totally new migrations to rename the tables from "roles" to "groups"