Fully finished group management page functionality. Next need to implement helper javascript functions for editing and deleting database fields asynchronously from the browswer. This will make future development much easier as the developer will only need to call a javascript method from a button to delete or edit any database field from the UI.
May consider implementing similar functionality for adding items to a database table.

Security still in place as on the backend we can set up rules for what users are allowed to modify what, and also user must be logged in regardless of rules (already implemented) in order for the javascript methods to have any effect.