Fixed an perplexing hang that was very obvious in large simulations --
"The root cause was processes would report statistics (via messages) back to
rank zero while it was shutting down (sending GVT updates to other processes)
-- still the rank zero's main loop was running which would silently eat
these statistics message.  When rank zero finally goes to process statistics
from all processes it would simply hang (because some of the messages were
lost earlier)."

The fix was to add an "MPI_Barrier" call at the end of the main simulation
loop so that all the processes wait and consistently exit their main 
simulation loop.  This ensures all simulation-phase messages are
correctly processed and all the processes enter finalization-phase 
in consistent state.

In addition, relaxed the order in which the statistics messages are
processed at rank 0 -- Earlier we used to print stats in a fixed order 
which *may* cause issues with many processes (when receive buffers get full
with many statistics messages from other processes).  So this loop processes
messages as and when they become available rather than in rank-order.