With simulations that have a very large range of time stamps the number of buckets in a rung of a ladder queue can become very large.
This was the case observed in waterfowl migration model where infective
waterfowl spread infection to neighboring flocks with timestamp of 0.125 and
when they migrate they schedule events with timestamp of 2592000000 (30 days
in milliseconds).  This wide spread causes many buckets to be created which
significantly increases the memory needs.

This check-in resolves this issue in the following 2 ways:

1. When events are moved into the ladder (from 'top' or 'bottom') the bucket
   with is adjusted such that no more than MAX_BUCKETS would be needed.  This
   of course is a deviation from the original ladder-queue paper

2. When a event is inserted and the bucket for the event is higher than 
   MAX_BUCKETS, the rung is rebuilt with a larger bucket size and events
   are redistributed.  This is also different that the original ladder-queue
   paper.

The operations don't seem to play a role in phold and pcs benchmarks because
the timestamps in these benchmarks are relatively dense/close.  These
operations only arise in simulations with a sparse/widespread time stamps.