Moved MUSE MPI calls from C++ bindings to standard C bindings -- The C++ bindings are being moved into BOOST-MPI starting with OpenMPI-2.0. Since OSC clusters have OpenMPI-2.0, the C++ bindings cannot be used there without Boost and I don't want to carry dependency on Boost.
This change was relatively straightforward and most of the changes
were in MPIHelper -- essentially changing MPI::Double to MPI_Double,
etc.  Some changes required short methods to be developed to minimize
changes in other parts of MUSE.  The use of the MPIHelper turned out
to be a good idea on the long run as it made this transition
relatively painless (took about 30 minutes to make changes, compile,
and test).

Also removed checks for constant exceptions in configure as we now use
custom exceptions.  Technically, with C bindings exceptions are not
relevant and they can eventually be totally cleaned-up in the
communicators.

The configure script also took a change to fix to streamline
overriding -O3 settings when debugging options are passed to configure
script.