Added code to spin-up and join threads in the MTBenchmark::run() method. Now MTBenchmark is actually doing some multi-threaded operations, albeit queueing data structures do not actually support multi-threading.
Minor modification to Benchmark.h to enable accessing numParitions so
that the corresponding number of threads are spun-up.  The
--partitions command-line argument is used to specify number of
partitions -- which is also used as the number of threads to run.

Modified the makefile to link with pthreads library and create a
symlink to mt_queue_bench to create the multi-threaded version of the
queueing benchmark.