Implemented naive fully locking thread safe event queue in the context of the MT queue benchmark.
Builds and runs, but the new queue seems to still have issues as it does not finish even when running sequentially, and returns the same error message as the non-thread safe structures when running with multiple threads. Will continue to look into it, not sure at the moment what could be wrong as all it does is take every method from the parent class and lock before calling the base function.