More changes to make changing colors a bit faster. The key issue was getting metadata via OpenLayer's API was slow (I think internally it was recreating arrays over-and-over causing API calls to be slow).
So EntityLayer now maintains arrays of metadata and features. Using the local
array makes the processing go a lot faster. The performance is now somewhat
acceptable but still sluggish. Need to use some setTimeout calls to batch-up
the entity creation and color updating so that the user can see the progress
and not wait long for the updates to happen.

However, for now it is tolerable.