A major change was a rehaul of the MySQL datbase to enable compression of text fields to try to reduce the storage overheads.
The best way to pick-up these migrations is to reseed the database and reload
the model. I have tried to make the model loading to go a bit faster by
directly implementing that logic in C++/sbfftool rather than through PHP.
Now, model loading is a bit faster, but not as fast as I would have liked it
(I am hoping for under 5 minutes to load model into the DB).

Some back-end work to get initial analysis entry added to the database.