Created a Lock Free Priority Queue based on work by Linden and Jonsson in 2013.
So far enqueue and dequeue work as expected single threaded. Next will be testing with multi threaded and validation testing to ensure queue performs well under high contention conditions. From there goal is to use this as the top tier of 3tHeap.

Please take a look and advise if you have any recommendations.