Added an initial, empty place holder class for heterogeneous-compute cpable parallel simulations. Of course, a lot more work is needed on this  class before it is really usable.
In this context, "heterogeneous-compute" refers to the use of CPU and GPU
for computation.  The GPU features of the simulation will be
enabled using OpenCL.  Here we have chosen to use OpenCL over
CUDA, because CUDA is limited to running only on nVidia platforms.
On the other hand, OpenCL (limited to Version 1.1) is supported on
all patforms.  We have consciously avoided using features from
OpenCL 2.0 or newer releases to ensure broad portability to nVidia
platforms.