Changes to the database and infrastructure to enable storing of snapshot data from multiple replications into the database. The snapshot table schema also took a change to include a sequence number to ease displaying the data.
Updated various links and associated functionality in the dashboard. The
only other link that is not yet operational is viewing the analysis/model
via the foresight map. This feature requires a bit more work on the backend
before it is going to be operational.

Further changes and enhancements to make the layout and organization of the 
main dashboard look nicer.