Experiment results with plots and brief discussion. Only thing not drafted at this point is the conclusion.
I'm hesitent to spend too much more time going into more details and wordsmithing without feedback first, as it's very possible the paper is off target and that work would be for nothing.

I'm hoping we can get this reviewed, edited, and defended within the next few weeks, as I have a very hard deadline of moving to California and starting my job in early-mid July.

Depending on when you'll be in Benton, potentially you could skim it and we could discuss next steps in person?

Thank you in advance, ~Matt.