Revised results and conclusion to represent final data and findings.
At this point, the results and conclusion is up to date with the most recent data and edited with my final findings, making it ready for feedback.

I have not yet gone through the edits previously prodivded by your annotated PDF (before the Results section), these will be incorporated tonight.

Please review the findings and discussion, it's fairly short compared to rest of the paper.

I have started reviewing the code from "USE", however it is pretty hard to interpret as it is all in C. I wanted to prioritize getting the results and conclusion sections to you to review, I had some trouble parsing the large amount of data. I will continue to try to find a good way to compare results from their system, at the moment I am aiming to try and match their "granularity" so the amount of extra processing time per event in microseconds is the same as our system and then compare relative speedup from adding threads. Does this sound like a good approach?

Thanks again,
~Matt