Adding benchmark bash script used to generate most of the data for my thesis.
Converting the output of this bash script into a csv format was done via a macro in my text editor, which is not ideal. I'm sure there was a better way to do this that would be preferred, in which case the bash script exists just to show the commands used to generate the data.

If at any point you want to actually used the bash scripts to collect data, you can reach out and I can explain my text conversion process, but I don't want to commit it and make it a standard practice because it is far from ideal.