Modified the laravel classes to account for the new 'kind' column added to the models database.
Now the 'Overview dashboard' page has tabs for chikv and zikv model kinds. 
The chikv tab displays the current Chikungunya models. Currently, the Zika
tab is empty (but has place holder text showing) because there are no Zika
models in the database. 

The next step is to add a preliminary Zika model with calibrations based on
data collected over summer