Updates to enable reuse of some of the files during simulation as well. To minimize dependencies during simulation, some of the features for XFig generation (that indirectly use gdal) have been conditionally complied in -- if the compiler flag NO_XFIG is defined, then Xfig generation is conditionally compiled out.
In addition, a feature change in model is to now write out the
rectangular population ring information for use during simulation.
The population ring latitude and longitude are written out in high
precision to minimize (it does not completely avoid and there are
subtle differences in the 9th decimal places in some cases) rounding
issues with double data type.

Added a PopRing class (that extends Ring) to enable reading of the
population ring information that is included in the model.  This class
is used by OSMData.  The PopRing is also used during simulation.