Check-in of changes to correctly record rollback statistics from parallel simulations. This feature is enabled only when the RB_STATS compiler-flag is enabled. Note that this is a compile-time feature. It does slow down the simulations and consequently should be disabled when not used. So by default it is disabled.
Changes to Agent and Scheduler to correctly track and report rollback stats
in the checkAndHandleRollback() method. Earlier it was incorrectly done in
the handleFutureAntiMessage method.  This check-in also includes changes
to Agent to report the number of states that were cancelled.  This is a measure
of unproductive/wasted work that was optimistically done by the simulation.