The current custom model generation does not work correctly in production -- the source of the issue is that the code currently assumes that the web-server and storage location are same -- however, in production that is not the case.
So the backed for custom model generation needs to be reworked so that it
operates correctly in production.

To facilitate reworking the backend, I am adding a test case than can be
run from the console. The test case has some fixed example inputs to mimic
operation of a web-browser submitting a request.  The test case is reatively
straightforward as all the core testing functionality is already provided
by Laravel.

Now that I have a test case I can easily test and troubleshoot this
functionality.