Corrected the "custom model generation" code path to correctly use storage locations that could be on a remote machine. Now the "custom model generation" uses SSH and Laravel's 'Storage' facade to access files and process the data.
In addtion, the sbfftool has slightly updated to accept a list of countries
from where the entities are to be extracted.  This streamlines the process
of generating custom models (and eliminates an extra step of having to merge
intermediate files and clean them)

This change also includes an additional feature for returning IDs of entities
that are adjacent to a given airport.  This feature is built around a shell
script that facilitates the process.  I will check-in the scripts in the next
commit (to keep check-ins focused on specific features).