A feature to be added to setting up analyses is to permit a set of infections to be pre-seeded around a given list of airports. This feature also enables injecting infections for analysis from 3rd party systems (trying to integrate with Foresight).
The process of extracting entities around airports is facilitated via
the `get_entities_near_airports.sh` shell script.  Internally, the shell
script uses sbfftool to extract the entities from a given model file.
The `ModelController` already includes a method to coordinate operations
via the shell script.  The script is executed over an SSH connection
so that the extraction is run on the storage machine (which is
different in production, than the machine on which the web-server
runs).

A couple of svn-ignore entries for the 'chikv' and 'zikv' directories
to avoid checking-in generated model files and parameters.