This commit includes the first phase of integrating with IBIS (https://ibis.eha.io). IBIS is a software system that uses Natural Language Processing (NLP) to predict potential epidemic threats, based on air travel.
Now, Foresight displays some of the pertinent information obtained from IBIS.
This information is displayed in the popup that appears when a user clicks
on an airport or an entity in a map.

Note: For this feature to correctly work, we need a login-id and password
to be setup on IBIS. The values are to be setup in the ".env" file in 
Foresight.