Added a RESTful API feature to obtain data fromn Foresight. The idea is that the REST-API will be utilized by 3rd party applications that wish to integrate with Foresight.
For this operation to work, the user's must supply an authetication token
from Foresight along with each API call. The authentication token is stored
in the Foresight database in the User table.  The authentiation token is
updated each time the user logs into the application.  This ensures that
the auth token changes somewhat frequently.  Currently, the API exposes
model and analysis snapshot information -- this is essentially the same
information that is displayed by Foresight (in map view), except the data
is obtained in JSON format via REST-API call.