Adding the primary interface directory and files constituting the PC2L framework.
This include/ directory meant to contain headers that provide external
facing interface to PC2L.  This directory is meant to contain only the
smallest subset of files that are necessary to compile an application
and link against the PC2L library.

In other words, best effort should be invested to ensure most of the
implementation lives inside the pc2l/src directory and is compiled
into the PC2L library.